+++ lib/src/parser/json_parser.dart
@@ -409,8 +409,9 @@ class JsonLocalizationParser {
     for (final e in value.entries) {
       if (!e.key.startsWith('@')) continue;
       // reserved translation structures
-      if (e.key == '@plural' || e.key == '@gender' || e.key == '@context')
+      if (e.key == '@plural' || e.key == '@gender' || e.key == '@context') {
         continue;
+      }
       // known doc fields
       if (e.key == '@description' ||
           e.key == '@example' ||